Eliminate babel config
