--- packages/babel-preset-cup/README.md
@@ -1,3 +1,0 @@
-# babel-preset-cup
-
-Babel preset for create-universal-package
