+++ packages/create-universal-package/lib/config/build-node.js
@@ -1,0 +1,14 @@
+const babel = require('rollup-plugin-babel');
+const {
+  getBabelConfig,
+  template,
+  formats,
+} = require('./build-shared');
+
+const plugins = [babel(getBabelConfig('node', [6]))];
+
+module.exports = {
+  plugins,
+  template,
+  formats,
+};
