--- packages/create-universal-package/lib/config/build.js
@@ -1,11 +1,0 @@
-const babel = require('rollup-plugin-babel');
-
-const plugins = [babel()];
-const template = ({env, format}) => `${env}.${format}.js`;
-const formats = ['es', 'cjs'];
-
-module.exports = {
-  plugins,
-  template,
-  formats,
-};
