+++ packages/create-universal-package/package.json
@@ -17,6 +17,8 @@
   },
   "dependencies": {
     "args": "^3.0.2",
+    "babel-plugin-transform-cup-globals": "^1.0.0-rc.3",
+    "babel-preset-env": "^1.6.0",
     "draftlog": "^1.0.12",
     "rollup": "^0.43.0",
     "rollup-plugin-alias": "^1.3.1",
