--- packages/fixture-package/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["cup"]
-}
